Move taskfields into its own component
