+++ src/components/task-fields.css
@@ -1,0 +1,16 @@
+@media (min-width: 768px) {
+  .task-fields {
+    column-count: 3;
+    column-gap: 40px;
+    margin-bottom: 10px;
+    column-rule: 1px solid #aaa;
+  }
+}
+
+.task-field {
+  column-count: 2;
+}
+
+.task-field__name {
+  color: #555;
+}
