+++ src/pages/task.css
@@ -17,20 +17,3 @@
     flex-grow: 1;
   }
 }
-
-@media (min-width: 768px) {
-  .task-fields {
-    column-count: 3;
-    column-gap: 40px;
-    margin-bottom: 10px;
-    column-rule: 1px solid #aaa;
-  }
-}
-
-.task-field {
-  column-count: 2;
-}
-
-.task-field__name {
-  color: #555;
-}
